Realigned title. Updated text content based on designs. Fixed scrolling issue on chat. Moved messages to its own component.
