+++ Sentiment Demo UI/src/app/components/info-panel/info-panel.component.html
@@ -1,8 +1,8 @@
 <section fxLayout="column" class="p-5 overflow-visible">
   <!-- Header -->
   <div class="mb-5">
-    <div class="mat-h1 m-0">Sentiment analysis</div>
-    <div class="mat-body-2 text-grey fw-normal mb-3">Real-time sentiment analysis demo</div>
+    <div class="mat-h1 m-0">Sentiment Analysis</div>
+    <div class="mat-body-2 text-grey fw-normal text-uppercase mb-3">Project template</div>
     <div class="mat-h4 fw-normal mb-3">Take the emotional pulse of any conversation with a ML sentiment analysis pipeline built in Quix.</div>
     <div fxLayout="row">
       <a mat-raised-button color="primary" class="small me-2" fxLayout="row" fxLayoutAlign="center center"
@@ -20,11 +20,11 @@
 
   <!-- Realtime messages feed -->
   <div fxLayout="column" fxLayoutAlign="start start" class="mb-5">
-    <div class="mat-h4 mb-2">Realtime messages feed</div>
+    <div class="mat-h4 mb-2">Real-time messages feed</div>
     <p class="mat-body-2 text-grey">
-      To feed our model, we <a [href]="'https://portal.platform.quix.ai/visualize/live/data?workspace=' + workspaceId + '&topic=messages'" target="_blank">ingest the messages</a> sent by our users from this very same UI. But you can 
-      <a [href]="'https://portal.platform.quix.ai/applications/ide/Sentiment%20Demo%20UI?workspace=' + workspaceId" target="_blank">customise the code</a> 
-      to connect any data source. For instance, try to select a Twitch chat from the following dropdown.
+      To feed the model, the project <a [href]="'https://portal.platform.quix.ai/visualize/live/data?workspace=' + workspaceId + '&topic=messages'" target="_blank">ingest the messages</a>
+      sent by users from this very UI. You can <a [href]="'https://portal.platform.quix.ai/applications/ide/Sentiment%20Demo%20UI?workspace=' + workspaceId" target="_blank">customise the code</a> 
+      to connect any data source (customer chats, social media, forums...). For instance, try to select a Twitch chat from the following dropdown.
     </p>
     <app-message-source-dropdown class="w-100"></app-message-source-dropdown>
     <button fxHide.gt-sm mat-flat-button class="mt-3 small" color="primary" (click)="scrollToChat()">
@@ -37,18 +37,21 @@
 
     <!-- Sentiment analysis -->
     <div fxLayout="column" class="mb-5">
-      <div class="mat-h4 mb-2">Sentiment Analysis</div>
-      <p class="mat-body-2 text-grey">Sentiment is <a [href]="'https://portal.platform.quix.ai/pipeline/deployments/0c40eef8-0b9e-46e6-965c-ea87637ee2e7?workspace=' + workspaceId" target="_blank">
-        calculated in real time</a> using a huggingface model.
-        Each message is assigned a value between -1 and 1. Then, we use a rolling windown (5 min) to analyse the trend.
+      <div class="mat-h4 mb-2">Sentiment analysis</div>
+      <p class="mat-body-2 text-grey">
+        Sentiment is <a [href]="'https://portal.platform.quix.ai/pipeline/deployments/0c40eef8-0b9e-46e6-965c-ea87637ee2e7?workspace=' + workspaceId" target="_blank">calculated in real time</a>
+        using a Hugging Face model. Each message is assigned a value between -1 and 1. Then, the service uses a rolling window to analyze the trend.
       </p>
       <app-sentiment-chart [style.height.px]="200"></app-sentiment-chart>
     </div>
 
     <!-- Interested? -->
     <div fxLayout="column">
-      <div class="mat-h4 mb-2">Are you intrigued?</div>
-      <p class="mat-body-2 text-grey mb-3">Try Quix for free to build your own pipeline. Or visit other demos to explore what else is possible.</p>
+      <div class="mat-h4 mb-2">Want to know more?</div>
+      <p class="mat-body-2 text-grey mb-3">
+        <a href="https://quix.io/docs/platform/tutorials/sentiment-analysis/index.html" target="_blank">Read the tutorial</a>
+        and try Quix for free to reproduce this project or build something similar. To discover what else is possible, see our other project templates.
+      </p>
       <div fxLayout="row">
         <a mat-raised-button color="primary" class="small me-2" href="https://portal.platform.quix.ai/self-sign-up" target="_blank">
           Sign up for free
@@ -57,7 +60,7 @@
         <a mat-raised-button color="accent" class="small" fxLayout="row" fxLayoutAlign="center center"
           href="https://github.com/orgs/quixio/repositories" target="_blank">
           <img src="assets/github.svg" alt="Quix" height="16" class="me-2" />
-          <span>Explore other demos</span>
+          <span>More project templates</span>
         </a>
       </div>
     </div>
