+++ Sentiment Demo UI/src/app/components/web-chat/web-chat.component.html
@@ -49,31 +49,7 @@
 
     <mat-card-content fxFlex fxLayout="column" class="bg-white mb-0 overflow-hidden">
       <div fxFlex class="overflow-auto"  #chatWrapper>
-        <div fxLayout="column" class="_message py-1 px-2 m-3 rounded-3 overflow-hidden" *ngFor="let message of messages">
-          <div fxFlex fxLayout="row" fxLayoutAlign="start start">
-            <div *ngIf="message.profilePic" [style.background-color]="message.profilePicColor" class="_profile-pic me-2">
-              <img [src]="message.profilePic" height="28">
-            </div>
-            <div fxLayout="column">
-              <p class="mb-0 fw-bold mat-body-2">{{ message.name }}</p>
-              <div class="_message-text">{{ message.value }}</div>
-            </div>
-          </div>
-          <div fxLayout="row" fxLayoutAlign="end center" class="mat-caption">
-            <p class="mb-0 me-2 text-grey-light mat-caption"> {{ getDateFromTimestamp(message.timestamp) | date:'hh:mm:ss' }}</p>
-            <div 
-              [class]="getColor(message?.sentiment!)" fxLayout="row" fxLayoutAlign="start center">
-              <mat-icon *ngIf="getSentimentIcon(message?.sentiment!) as icon" inline class="material-icons-outlined me-1" [style.fontSize.px]="16">
-                {{ icon }}
-              </mat-icon>
-              <div *ngIf="!message.sentiment" fxLayoutAlign="start center">
-                <mat-spinner diameter="12" class="text-grey-light"></mat-spinner>
-                <p class="ms-1 mb-0">...</p>
-              </div>
-              <p *ngIf="message.sentiment" class="mb-0 fw-bold">{{  (message.sentiment! > 0 ? '+' : '') + (message.sentiment | number : '1.2-2') }}</p>
-            </div>
-          </div>
-        </div>        
+        <app-message *ngFor="let message of messages" [message]="message"></app-message> 
       </div>
 
       <!-- Standard Chat isTyping sentiment indicator -->
