+++ Sentiment Demo UI/src/app/components/web-chat/web-chat.component.ts
@@ -11,8 +11,8 @@ import { ShareChatroomDialogComponent } from '../dialogs/share-chatroom-dialog/s
 import { RoomService } from 'src/app/services/room.service';
 import { Colors } from 'src/app/constants/colors';
 
-const POSITIVE_THRESHOLD = 0.5;
-const NEGATIVE_THRESHOLD = -0.5;
+export const POSITIVE_THRESHOLD = 0.5;
+export const NEGATIVE_THRESHOLD = -0.5;
 
 export class UserTyping {
   timeout?: Subscription;
@@ -31,6 +31,7 @@ export class WebChatComponent implements OnInit {
   profilePicColor: string;
 
   isTwitch: boolean = true;
+  scrollTimeoutId: any;
 
   @ViewChild('chatWrapper') chatWrapperEle: ElementRef<HTMLElement>;
   @ViewChild('messageInput') messageInputEle: ElementRef;
@@ -239,78 +240,31 @@ export class WebChatComponent implements OnInit {
    * 
    * @returns The Html message.
    */
-    public getTypingMessage(): string | undefined {  
+  public getTypingMessage(): string | undefined {  
 
-      const users = Array.from(this.usersTyping.entries())
-      .map(([key, value]) => ({
-        name: this.titleCasePipe.transform(key),
-        sentiment: value.sentiment!,
-      }));
+    const users = Array.from(this.usersTyping.entries())
+    .map(([key, value]) => ({
+      name: this.titleCasePipe.transform(key),
+      sentiment: value.sentiment!,
+    }));
 
-      if (!users.length) return undefined;
+    if (!users.length) return undefined;
 
-      if (users.length === 1) {
-        const [user] = users;
-        return `<b>${user.name}</b> is typing...`;
-      }
-
-      if (users.length < 3) {
-        const usersJoined = users.map((m) => m.name).join(' and ');
-        return `<b>${usersJoined}</b> are typing...`; 
-      }
-
-      if (users.length >= 3) {
-        return `<b>${users.length}</b> users are typing...`;    
-      }
-
-      return undefined
+    if (users.length === 1) {
+      const [user] = users;
+      return `<b>${user.name}</b> is typing...`;
     }
 
-  /**
-   * Takes the sentiment value of a message and returns the
-   * appropriate color to be rendered in the template.
-   * 
-   * @param sentiment The sentiment of the message.
-   * @returns The Html class.
-   */
-  public getColor(sentiment: number): string {
-    if (sentiment > POSITIVE_THRESHOLD) {
-      return 'text-success';
-    } else if (sentiment < NEGATIVE_THRESHOLD) {
-      return 'text-danger';
-    } 
-      
-    return 'text-grey-light';
-  }
-
-  /**
-   * Takes the sentiment value of a message and returns the
-   * appropriate icon to be rendered in the template.
-   * 
-   * @param sentiment The sentiment of the message.
-   * @returns The icon name.
-   */
-  public getSentimentIcon(sentiment: number): string | undefined {
-    if (sentiment > POSITIVE_THRESHOLD) {
-      return 'sentiment_satisfied';
-    } else if (sentiment < NEGATIVE_THRESHOLD) {
-      return 'sentiment_dissatisfied';
-    } else if (sentiment === 0) {
-      return 'sentiment_neutral';
+    if (users.length < 3) {
+      const usersJoined = users.map((m) => m.name).join(' and ');
+      return `<b>${usersJoined}</b> are typing...`; 
     }
-      
-    return undefined;
-  }
 
-  /**
-   * Takes the date timestamp, divides it by 1000000
-   * and then creates a JS date from it to be used in the template.
-   * 
-   * @param timestamp The date timestamp.
-   * @returns The new Date. 
-   */
-  public getDateFromTimestamp(timestamp: number) {
-    return new Date(timestamp / 1000000)
+    if (users.length >= 3) {
+      return `<b>${users.length}</b> users are typing...`;    
+    }
+
+    return undefined
   }
 
   /**
@@ -363,12 +317,15 @@ export class WebChatComponent implements OnInit {
    * Util method to scroll the user to the bottom of the chat
    */
   scrollToChatBottom(isCheckBottom?: boolean): void {
+    const threshold = 30;
     const el = this.chatWrapperEle.nativeElement;
-    const isScrollToBottom = el.offsetHeight + el.scrollTop >= el.scrollHeight;
+    const isScrollToBottom = Math.round(el.scrollTop + el.clientHeight) >= el.scrollHeight - threshold;
 
     if (isCheckBottom && !isScrollToBottom) return;
-    
-    setTimeout(() => (el.scrollTop = el.scrollHeight));
+  
+    // Clear the existing timeout if it exists
+    if (this.scrollTimeoutId) clearTimeout(this.scrollTimeoutId);
+    this.scrollTimeoutId = setTimeout(() => el.scrollTop = el.scrollHeight, 0);
   }
 
   openShareChatroomDialog(): void {
